requirements.txt
python-dotenv==1.0.1
matplotlib==3.10.0
aiogram==3.17.0
